Added validation

+++ iv_backend/vault/views.py
@@ -106,9 +106,8 @@ class VaultViewSet(viewsets.ModelViewSet):
             if not TeamMembership.objects.filter(
                 user=self.request.user, team=team, role=TeamMembership.ADMIN
             ).exists():
-                raise PermissionDenied(
-                    "Only team admins can create vaults for the team."
-                )
+                    return Response({'error': 'Only team admins can create team vaults.'}, status=status.HTTP_403_FORBIDDEN)
+
 
             serializer.save(owner=self.request.user, team=team)
         else:
@@ -123,7 +122,10 @@ class VaultViewSet(viewsets.ModelViewSet):
             if not TeamMembership.objects.filter(
                 user=self.request.user, team=instance.team, role=TeamMembership.ADMIN
             ).exists():
-                raise PermissionDenied("Only team admins can update team vaults.")
+                return Response({'error': 'Only team admins can update team vaults.'}, status=status.HTTP_403_FORBIDDEN)
+            
+        if instance.owner != self.request.user:
+            return Response({'error': 'You do not have permission to modify this vault.'}, status=status.HTTP_403_FORBIDDEN)
 
         serializer.save()
 
@@ -131,10 +133,11 @@ class VaultViewSet(viewsets.ModelViewSet):
         # Check if the vault is associated with a team
         if instance.team:
             # Ensure the user is an admin of the team
-            if not TeamMembership.objects.filter(
-                user=self.request.user, team=instance.team, role=TeamMembership.ADMIN
-            ).exists():
-                raise PermissionDenied("Only team admins can delete team vaults.")
+            if not TeamMembership.objects.filter(user=self.request.user, team=instance.team, role=TeamMembership.ADMIN).exists():
+                return Response({'error': 'Only team admins can delete team vaults.'}, status=status.HTTP_403_FORBIDDEN)
+        
+        if instance.owner != self.request.user:
+            return Response({'error': 'You do not have permission to modify this vault.'}, status=status.HTTP_403_FORBIDDEN)
 
         instance.delete()
